+++ PowerEditor/src/WinControls/VerticalFileSwitcher/VerticalFileSwitcherListView.h
@@ -29,7 +29,7 @@
 #ifndef VERTICALFILESWITCHERLISTVIEW_H
 #define VERTICALFILESWITCHERLISTVIEW_H
 
-#include "window.h"
+#include "Window.h"
 #include "TaskListDlg.h"
 
 class Buffer;
@@ -69,7 +69,7 @@ public:
 		ListView_DeleteColumn(_hSelf, i);
 	};
 	int nbSelectedFiles() const {
-		return SendMessage(_hSelf, LVM_GETSELECTEDCOUNT, 0, 0);
+		return static_cast<int32_t>(SendMessage(_hSelf, LVM_GETSELECTEDCOUNT, 0, 0));
 	};
 
 	std::vector<SwitcherFileInfo> getSelectedFiles(bool reverse = false) const;
